Distance from camera works
Able to find the horizontal distance from camera to reflective tape based on distance between the contour and angle of the robot. We need to change height in order for the program to work based on where the camera is. The focal length may need to change because the calculation returns a value within 1 inch of actual.